ai/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
filelock==3.20.2
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.4
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-text-splitters==1.1.0
langsmith==0.5.2
lxml==6.0.2
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.0
orjson==3.11.5
packaging==25.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pypdf==6.5.0
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
shellingham==1.5.4
SQLAlchemy==2.0.45
tenacity==9.1.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
yarl==1.22.0
zstandard==0.25.0

# Speech-to-Text (STT) dependencies
openai-whisper==20231117
ffmpeg-python==0.2.0
